ml-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
redis==5.0.1
confluent-kafka==2.3.0
python-dotenv==1.2.2
sentence-transformers==2.7.0
transformers==4.40.0
prometheus-fastapi-instrumentator==6.1.0